Scanner Settings on Layer Load
Got the scanner settings to update properly when a new scanning layer is loaded.  Most of my difficulties have been related to figuring out how to talk to the scanner object, since it's established in index.js (a convention I lifted from the Dynamasoft demos).  Remedy at this stage is to store everything the  Scanner might interact with in some sort of DOM element.  Scanner can be updated/interacted with by simulating events on these elements..