Complete Update to Pct Based Templates
Finished switching over to percentage based template definition, including most of the logic.  Trying to move away from referencing the HTML input elements directly, and instead relying on the parameters attached to the ScanningLayer.  Need to be able to adjust grid size in ("DevMode"), but okay with requiring that these params are manually entered into the SL template instead of dynamically updating the ScanningLayer