EOD Commit
Added some functionality for a "DevMode".

MORE IMPORTANTLY, it looks like camera toggling is only an issue when working from the local file.  It works if we run via a localhost.  Was able to do this in VS code with the Live Server extension, making me hopeful that it won't be an issue when we deploy.