Symphony Scan Demo
Updated settings to reflect 4K camera.  This resolution is still hard coded in places, and I am not confident in dynamically changing resolution.  This commit demonstrated the capability to read 24 Simport tubes simultaneously.  Backing up here because I'm going to begin trying to port to GAS and there is non-zero chance I break something in the process